pyproject.toml
[project]
name = "antlr4-tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
    "antlr4-tools>=0.2.2",
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.1",
    "flagembedding>=1.3.5",
    "numpy>=2.3.5",
    "qdrant-client>=1.16.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
]
